WIP: price estimation refactor
