+++ src/aleph/schemas/pending_messages.py
@@ -17,9 +17,11 @@ TODO: this module should reasonably be part of aleph message, if only
       in aleph-client.
 """
 
+import datetime as dt
 from typing import Any, Dict, Generic, Literal, Type
 
 from aleph_message.models import (
+    Chain,
     AggregateContent,
     ForgetContent,
     InstanceContent,
@@ -29,9 +31,10 @@ from aleph_message.models import (
     ProgramContent,
     StoreContent,
 )
-from pydantic import ValidationError, root_validator
+from pydantic import ValidationError, root_validator, validator
 
 import aleph.toolkit.json as aleph_json
+from aleph.toolkit.timestamp import timestamp_to_datetime
 from aleph.exceptions import UnknownHashError
 from aleph.schemas.base_messages import AlephBaseMessage, ContentType, MType
 from aleph.types.message_status import InvalidMessageFormat
@@ -40,56 +43,82 @@ from aleph.utils import item_type_from_hash
 MAX_INLINE_SIZE = 200000  # 200kb max inline content size.
 
 
+def base_pending_message_load_content(values):
+    """
+    Preload inline content. We let the CCN populate this field later
+    on for ipfs and storage item types.
+
+    Sets the default value for item_type, if required.
+    """
+
+    item_hash = values.get("item_hash")
+    if item_hash is None:
+        raise ValueError("Could not determine item hash")
+    item_content = values.get("item_content")
+
+    try:
+        default_item_type = (
+            item_type_from_hash(item_hash)
+            if item_content is None
+            else ItemType.inline
+        )
+    except UnknownHashError:
+        raise ValueError(f"Unexpected hash type: '{item_hash}'")
+
+    input_item_type = values.get("item_type")
+    item_type = input_item_type or default_item_type
+
+    if item_type == ItemType.inline:
+        if item_content is None:
+            raise ValueError("Item content not specified for inline item type")
+
+        if len(item_content) > MAX_INLINE_SIZE:
+            raise ValueError("Message too long")
+        try:
+            values["content"] = aleph_json.loads(item_content)
+        except aleph_json.DecodeError as e:
+            raise ValueError("Message content is not valid JSON data") from e
+    else:
+        if item_content is not None:
+            raise ValueError(f"{item_type} messages cannot define item_content")
+
+    # Store back the default item_type if not specified
+    if input_item_type is None:
+        values["item_type"] = default_item_type.value
+
+    return values
+
+
+def base_pending_message_validator_check_time(v, values):
+    """
+    Parses the time field as a UTC datetime. Contrary to the default datetime
+    validator, this implementation raises an exception if the time field is
+    too far in the future.
+    """
+
+    if isinstance(v, dt.datetime):
+        return v
+
+    return timestamp_to_datetime(v)
+
+
 class BasePendingMessage(AlephBaseMessage, Generic[MType, ContentType]):
     """
     A raw Aleph message, as sent by users to the Aleph network.
     """
 
+    sender: str
+    chain: Chain
+    type: MType 
+    time: dt.datetime
+
     @root_validator(pre=True)
     def load_content(cls, values):
-        """
-        Preload inline content. We let the CCN populate this field later
-        on for ipfs and storage item types.
-
-        Sets the default value for item_type, if required.
-        """
-
-        item_hash = values.get("item_hash")
-        if item_hash is None:
-            raise ValueError("Could not determine item hash")
-        item_content = values.get("item_content")
-
-        try:
-            default_item_type = (
-                item_type_from_hash(item_hash)
-                if item_content is None
-                else ItemType.inline
-            )
-        except UnknownHashError:
-            raise ValueError(f"Unexpected hash type: '{item_hash}'")
-
-        input_item_type = values.get("item_type")
-        item_type = input_item_type or default_item_type
-
-        if item_type == ItemType.inline:
-            if item_content is None:
-                raise ValueError("Item content not specified for inline item type")
-
-            if len(item_content) > MAX_INLINE_SIZE:
-                raise ValueError("Message too long")
-            try:
-                values["content"] = aleph_json.loads(item_content)
-            except aleph_json.DecodeError as e:
-                raise ValueError("Message content is not valid JSON data") from e
-        else:
-            if item_content is not None:
-                raise ValueError(f"{item_type} messages cannot define item_content")
-
-        # Store back the default item_type if not specified
-        if input_item_type is None:
-            values["item_type"] = default_item_type.value
-
-        return values
+       return base_pending_message_load_content(values)
+    
+    @validator("time", pre=True)
+    def check_time(cls, v, values):
+        return base_pending_message_validator_check_time(v, values)
 
 
 class PendingAggregateMessage(
