+++ src/aleph/web/controllers/prices.py
@@ -1,4 +1,3 @@
-import datetime as dt
 import logging
 from dataclasses import dataclass
 from decimal import Decimal
@@ -13,7 +12,6 @@ from pydantic import BaseModel, Field
 import aleph.toolkit.json as aleph_json
 from aleph.db.accessors.messages import get_message_by_item_hash, get_message_status
 from aleph.db.models import MessageDb
-from aleph.db.models.pending_messages import PendingMessageDb
 from aleph.schemas.api.costs import EstimatedCostsResponse
 from aleph.services.cost import (
     get_payment_type,
@@ -21,6 +19,7 @@ from aleph.services.cost import (
     get_total_and_detailed_costs_from_db,
 )
 from aleph.toolkit.costs import format_cost_str
+from aleph.schemas.cost_estimation_messages import validate_cost_estimation_message_content, validate_cost_estimation_message_dict
 from aleph.types.db_session import DbSession
 from aleph.types.message_status import MessageStatus
 from aleph.web.controllers.app_state_getters import (
@@ -122,7 +121,6 @@ async def message_price(request: web.Request):
 
 
 class PubMessageRequest(BaseModel):
-    sync: bool = False
     message_dict: Dict[str, Any] = Field(alias="message")
 
 
@@ -134,23 +132,10 @@ async def message_price_estimate(request: web.Request):
 
     with session_factory() as session:
         parsed_body = PubMessageRequest.parse_obj(await request.json())
-        pending_message = PendingMessageDb.from_message_dict(
-            parsed_body.message_dict,
-            reception_time=dt.datetime.now(),
-            fetched=False,
-        )
-
-        content_dict = await storage_service.get_message_content(pending_message)
-
-        message = MessageDb.from_pending_message(
-            pending_message,
-            content_dict=content_dict.value,
-            content_size=len(content_dict.raw_value),
-        )
-
+        message = validate_cost_estimation_message_dict(parsed_body.message_dict)
+        content = await validate_cost_estimation_message_content(message, storage_service)
         item_hash = message.item_hash
-        content: ExecutableContent = message.parsed_content
-
+ 
         try:
             payment_type = get_payment_type(content)
             required_tokens, costs = get_total_and_detailed_costs(
